// src/proto/dapr/proto/durabletask/v1/orchestration_pb.d.ts
export enum OrchestrationStatus {
  /**
   * @generated from enum value: ORCHESTRATION_STATUS_RUNNING = 0;
   */
  RUNNING = 0,

  /**
   * @generated from enum value: ORCHESTRATION_STATUS_COMPLETED = 1;
   */
  COMPLETED = 1,

  /**
   * @generated from enum value: ORCHESTRATION_STATUS_CONTINUED_AS_NEW = 2;
   */
  CONTINUED_AS_NEW = 2,

  /**
   * @generated from enum value: ORCHESTRATION_STATUS_FAILED = 3;
   */
  FAILED = 3,

  /**
   * @generated from enum value: ORCHESTRATION_STATUS_CANCELED = 4;
   */
  CANCELED = 4,

  /**
   * @generated from enum value: ORCHESTRATION_STATUS_TERMINATED = 5;
   */
  TERMINATED = 5,

  /**
   * @generated from enum value: ORCHESTRATION_STATUS_PENDING = 6;
   */
  PENDING = 6,

  /**
   * @generated from enum value: ORCHESTRATION_STATUS_SUSPENDED = 7;
   */
  SUSPENDED = 7,

  /**
   * @generated from enum value: ORCHESTRATION_STATUS_STALLED = 8;
   */
  STALLED = 8,
}